+++ Client/riepOrder.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script> 
-    <script type="module" src="main.js" defer></script>
+    <script type="module" src="riepOrderApp.js" defer></script>
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
       rel="stylesheet"
@@ -14,7 +14,7 @@
     <title>Riepilogo Ordine</title>
   </head>
   <body>
-    <main
+    <div
       class="container d-flex flex-column p-2 m-5 align-items-center justify-content-center"
       id="card-riep-order" 
     >
@@ -28,7 +28,7 @@
         </div>
         <button type="button" class="btn btn-danger" id="confirm-order-btn">CONFIRM ORDER</button>
       </div>
-    </main>
+    </div>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
